refactor: remove SidebarMenuNavigationItem and inline code
